Export shared functions and move into their own file, add remote model fetch
