+++ package.json
@@ -2,6 +2,7 @@
   "type": "module",
   "main": "src/headless.js",
   "dependencies": {
+    "axios": "^1.4.0",
     "gl": "^6.0.2",
     "jsdom": "^22.1.0",
     "node-fetch": "^3.3.1",
